requirement.txt
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.9.197
botocore==1.12.197
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.1.1
gensim==3.8.0
idna==2.8
itsdangerous==1.1.0
jaconv==0.2.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
knp-utils==0.42
MarkupSafe==1.1.1
mecab-python3==0.996.2
more-itertools==7.2.0
numpy==1.17.0
pandas==0.25.0
pexpect==4.7.0
ptyprocess==0.6.0
pyknp==0.4.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
soupsieve==1.9.2
urllib3==1.25.3
Werkzeug==0.15.5
